requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.16.1
huggingface_hub
google-generativeai
pillow
numpy